refactor: prepare for mayan route cleanup
moves index to another file for better future diff
